requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
tweepy
transformers
torch
torchvision
torchaudio
sentencepiece
pytesseract
requests
Pillow
python-telegram-bot